Fix killing of processes

+++ utils/utils.js
@@ -67,7 +67,8 @@ function execPromise (command, cwd, logger = null) {
 }
 
 const killProcess = async (pid) => {
-  if (!pid) return
+  if (!pid || !Number.isInteger(pid)) return
+  console.log('kill')
   if (/^win/.test(process.platform)) {
     return execPromise(`taskkill /F /pid ${pid}`)
   }
